% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarterLake}
\alias{CarterLake}
\title{Carter Lake point cloud data}
\format{
## 'las'
An LAS dataset with 5.25 million points
}
\source{
<https://coloradohazardmapping.com/lidarDownload>
}
\usage{
data(CarterLake
)
}
\description{
Lidar point cloud from Carter Lake, Colorado
}
\keyword{datasets}
