% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_clean.R
\name{pc_clean}
\alias{pc_clean}
\title{Clean up point clouds with lots of noise}
\usage{
pc_clean(las, denoise = TRUE, ground = TRUE)
}
\arguments{
\item{las}{A point cloud in .las or .laz format}

\item{denoise}{Boolean. Set to "TRUE" to remove erroneous points}

\item{ground}{Boolean. Set to "TRUE" to classify ground points}
}
\value{
The colorized point cloud.
}
\description{
Classify noise and remove along with points flagged "Withheld" from a point
cloud and classify ground if needed. The idea behind this function is to get
the point cloud ready for further processing for visualization.
}
\examples{
# Example code
color_las = pc_color(las=grey_las, ortho=ortho)
}
