% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_segment.R
\name{pc_segment}
\alias{pc_segment}
\title{Segment trees in a point cloud.}
\usage{
pc_segment(las, chm, ttops)
}
\arguments{
\item{las}{A point cloud in .las or .laz format}

\item{chm}{Spatraster canopy height model.}

\item{ttops}{SFC MULTIPOINT object with tree locations like what comes out of
lidR::locate_trees().}
}
\value{
The segmented point cloud with additional height-above-ground info.
}
\description{
Segments trees in a point cloud and adds a treeID field. This function
normalizes the point cloud by height prior to tree segmentation then
unnormalizes heights to return a point cloud with the original Z data. It
saves ground-normalized point heights in a field called "normHt." I can't
remember why I thought this was useful. Maybe to remove points below a
certain height for better visualization or something.
}
\examples{
# Example code
segmented_pc = pc_segment(las, chm, ttops)
}
